+++ src/main/java/com/cdkj/ylq/bo/IOverdueBO.java
@@ -1,5 +1,6 @@
 package com.cdkj.ylq.bo;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.cdkj.ylq.bo.base.IPaginableBO;
@@ -8,7 +9,7 @@ import com.cdkj.ylq.domain.Overdue;
 public interface IOverdueBO extends IPaginableBO<Overdue> {
 
     public void saveOverdue(String userId, String borrowCode, Integer days,
-            Long amount, String result);
+            BigDecimal amount, String result);
 
     public List<Overdue> queryOverdueList(Overdue condition);
